Add possibility to build all examples at once
Add prepare.sh script to configure the build.

The prepare.sh script will also source the ARMGCC_DIR variable, which is
set in the armgcc_version file.

The chosen_board variable from prepare.sh will be used in uppercase for
the define BOARD_{chosen_board} that will be used in the examples to set
the board specific periphery.

Also add .gitignore