Add PicoCoreMX8MP support
Adapt the examples to support the PicoCoreMX8MP.
Most of the board specific configuration is done in board.h and
pin_mux.c but sometimes the main code has to be adapted too.

Also fix the clock init of the audio clock in clock_config.c for
examples that dont use audio.
The examples could not be started in linux, because it got stuck in the
while loop, while waiting for the AUDIOMIX to get ready.

Now only examples that define "USE_RPMSG_AUDIO" will activate the AUDIO
clocks.